requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
plotly>=5.17.0
pdfplumber>=0.9.0
pillow>=10.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
yfinance>=0.2.28
requests>=2.31.0
urllib3>=2.0.0
markdown>=3.4.0